Restructure database

+++ database/migrations/2022_01_15_222153_create_customer_debt_payments_table.php
@@ -21,7 +21,7 @@ class CreateCustomerDebtPaymentsTable extends Migration
             $table->double('amount_paid');
             $table->double('balance');
             $table->date('date');
-            $table->integer('created_by')->unsigned();
+             $table->unsignedBigInteger('created_by')->unsigned();
             $table->timestamps();
         });
