+++ database/migrations/2023_01_03_223917_create_reservation_invoices_table.php
@@ -23,10 +23,10 @@ class CreateReservationInvoicesTable extends Migration
             $table->decimal('tax',8 , 2);
             $table->decimal('total_amount', 8, 2)->storedAs('amount + tax');
             $table->timestamp('issued_on')->default(DB::raw('CURRENT_TIMESTAMP'));
-            $table->integer('issued_by')->unsigned();
+            $table->unsignedBigInteger('issued_by')->unsigned();
             $table->enum('status', ['pending', 'paid', 'cancelled'])->default('pending');
             $table->string('payment_method')->nullable();
-            $table->integer('cancelled_by')->unsigned()->nullable();
+            $table->unsignedBigInteger('cancelled_by')->unsigned()->nullable();
             $table->timestamp('paid_on')->nullable();
             $table->timestamp('cancelled_on')->nullable();
             $table->timestamps();
